Add integration test for plugin
