+++ integration_test/spec/postgresql_spec.rb
@@ -14,7 +14,7 @@ context "PostgreSQL (AR#{ar_version})" do
     Arproxy.configure do |config|
       config.adapter = 'postgresql'
       config.use HelloProxy
-      config.use QueryLogger
+      config.plugin :query_logger
     end
     Arproxy.enable!
   end
