+++ integration_test/spec/shared_examples/custom_proxies.rb
@@ -1,25 +1,6 @@
 class Product < ActiveRecord::Base
 end
 
-class QueryLogger < Arproxy::Base
-  def execute(sql, name = nil)
-    @@log ||= []
-    @@log << sql
-    if ENV['DEBUG']
-      puts "QueryLogger: [#{name}] #{sql}"
-    end
-    super
-  end
-
-  def self.log
-    @@log
-  end
-
-  def self.reset!
-    @@log = []
-  end
-end
-
 class HelloProxy < Arproxy::Base
   def execute(sql, name = nil)
     super("#{sql} -- Hello Arproxy!", name)
