+++ integration_test/spec/sqlite3_spec.rb
@@ -10,7 +10,7 @@ context "SQLite3 (AR#{ar_version})" do
     Arproxy.configure do |config|
       config.adapter = 'sqlite3'
       config.use HelloProxy
-      config.use QueryLogger
+      config.plugin :query_logger
     end
     Arproxy.enable!
   end
